Update libraries/ManiaLivePlugins/MLEPP/IRC/IRC.php
Added public function to let other plugins announce to IRC. Either to all connected channels or specific channel.
Tested it out and seems to work just fine. Really should be expanded to allow for channel array to be passed to it. But to single channel or all it works just fine.
It does very basic errorreporting in console to for dev's.